Add the "obtainability analyzer," which lists unobtainable items.
The `recipes.json` in particular is a list of handwritten "recipes" for
teaching the analyzer how to obtain certain items from other ones. It's
also a list of things the grist cost generator doesn't understand (such
as smithing upgrade recipes). Future work could be reducing the size of
this list by making more interpreters.